Fixed some stuff

+++ src/main/java/sk/ness/academy/dto/AuthorStats.java
@@ -2,13 +2,13 @@ package sk.ness.academy.dto;
 
 public class AuthorStats {
 
-	public AuthorStats(String authorName, Integer articleCount) {
+	public AuthorStats(String authorName, Long articleCount) {
 		this.authorName = authorName;
 		this.articleCount = articleCount;
 	}
 
 	  private String authorName;
-	  private Integer articleCount;
+	  private Long articleCount;
 
 	  public String getAuthorName() {
 	    return this.authorName;
@@ -18,11 +18,11 @@ public class AuthorStats {
 	    this.authorName = authorName;
 	  }
 
-	  public Integer getArticleCount() {
+	  public Long getArticleCount() {
 	    return this.articleCount;
 	  }
 
-	  public void setArticleCount(final Integer articleCount) {
+	  public void setArticleCount(final Long articleCount) {
 	    this.articleCount = articleCount;
 	  }
